changes in adminMatriculacioController.php, AsignaturasCrudController.php, EgresadosAdminController, Asignaturas(Models), index.blade(inscriptos), index.blade.(cursadas), index.blade(profesores)

+++ app/Http/Controllers/Admin/AsignaturasCrudController.php
@@ -123,14 +123,34 @@ class AsignaturasCrudController extends BaseController
 
         }
 
-        // Si no tiene carreras vinculadas, procedemos a eliminarla
+        // Verificar si la asignatura tiene relaciones con cursadas
+        if($asignatura->cursadas()->exists()) {
+            return redirect()->route('admin.asignaturas.index')
+                ->with('error', 'No se pudo eliminar la asignatura porque tiene cursadas asociadas.');
+        }
+
+        //verificar si la asignatura tiene relaciones con mesas
+        if($asignatura->mesas()->exists()) {
+            return redirect()->route('admin.asignaturas.index')
+                ->with('error', 'No se pudo eliminar la asignatura porque tiene mesas asociadas.');
+        }
+
+        //verificar si la asignatura tiene relaciones con examenes
+        if($asignatura->examenes()->exists()) {
+            return redirect()->route('admin.asignaturas.index')
+                ->with('error', 'No se pudo eliminar la asignatura porque tiene examenes asociados.');
+        }
+
+        // Si no tiene relaciones, procedemos a eliminarla
         $asignatura->delete();
+
+
 
         return redirect()->route('admin.asignaturas.index')
             ->with('mensaje', 'Se ha eliminado la asignatura');
     } catch (\Exception $e) {
         return redirect()->route('admin.asignaturas.index')
-            ->with('error', 'No se pudo eliminar la asignatura. Error: ' . $e->getMessage());
+            ->with('error', 'No se pudo eliminar la asignatura'. $e->getMessage());
     }
 }
